+++ app/_document.tsx
@@ -8,21 +8,7 @@ class MyDocument extends Document {
           {/* Add the canonical link tag */}
           <link rel="canonical" href="https://aizaz-khalid-portfolio-website.vercel.app/" />
 
-           {/* Google Analytics */}
-          <script
-            async
-            src="https://www.googletagmanager.com/gtag/js?id=G-6BXRP55Z4E"
-          ></script>
-          <script
-            dangerouslySetInnerHTML={{
-              __html: `
-                window.dataLayer = window.dataLayer || [];
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-                gtag('config', 'G-6BXRP55Z4E');
-              `,
-            }}
-          />
+           
 
 
         </Head>
